cc: A pixel test for readbacks of layers with CanClipSelf == true.
TextureLayer is a self clipping layer, and this test verifies that
we are able to correctly copy the entire layer when it is outside
of the viewport.

Also includes a couple unit tests for TextureMailboxes on layers and
deleting the layer from the tree.

R=piman
BUG=263512

Review URL: https://chromiumcodereview.appspot.com/20048004