Correct bit-shift in IPC hash generation.
The actual bit shifting that happens does not match what is specified in the
comment. With the current implementation we will shift the PID over 14 bits,
this sets the top 22 bits with the bottom 22 bits of the PID.  When then OR
the count with 0x3fff giving us 14 bits which we shift over 8 bits. This will
end up OR'ing 8 bits of the count with 8 bits of the PID.

Instead, we shift the PID over 22 bits, giving is 10 bits of the PID and shift
the 14 bits of the count over 8. There should be no OR'ing between the PID and
the count. This then matches what is specified in the comments.

BUG=

Review URL: https://chromiumcodereview.appspot.com/20106002