Properly deprecate the IdleAction policy.
An earlier change had re-purposed it to IdleActionAC, which is not going
to work well in terms of backward-compatibility when multiple Chrome OS
versions / old configuration data is present. This change assigns a new
ID to IdleActionAC and makes IdleActionAC and IdleActionBattery to the
deprecated IdleAction policy.

BUG=chromium:254404
TEST=Manual: configure the old IdleAction policy and make sure it's respected.

Review URL: https://chromiumcodereview.appspot.com/20152005