DevTools: send trace events as bare object, not stringized JSON object
This makes reporting of trace events consistent with the rest
of devtools protocol: we now send trace events as part of the
notification JSON object, not encapsulated into a JSON
string. This changes the protocol, but telemetry
appears to be flexible enough to digest this without any code change.

R=pfeldman@chromium.org

Review URL: https://codereview.chromium.org/18474006