rAc: Allow the wallet interstitial to set the size of the dialog.
I had strayed away from this approach because I wanted to let the dialog stay the same size when the interstitial came up (to avoid jarring resizes). But that creates a very funky looking dialog in many cases (and results in a ton of bug reports in the dogfoods).

This new patch has the very pleasant side effect that the card image stays in the same place between the first ("generating...") and second ("generated/Okay, got it") dialog.

BUG=260415

Review URL: https://chromiumcodereview.appspot.com/20169004