Restructure Android app's ChromotingJniInstance::ProvideSecret()
This was done to make the timeframe for construction of the std::string wrapping the user's PIN clearer. It should fix a threading issue where the underlying const char* containing the PIN was being destroyed before being wrapped, resulting in the transmission of a bogus PIN to the host and a subsequent authentication rejection.

Review URL: https://chromiumcodereview.appspot.com/20226002