Call glTexImage2D() on a 1x1 pixel instead of a 0x0 NULL pixel.
Calling glTexImage2D() on a 0x0 pixel with a NULL buffer doesn't cause the storage on the texture to get redefined and as such the previously bound GpuMemoryBuffer remains bound to the texture. This causes undefined behavior (such as hangs) when we try to map the bound buffer.

glTexImage2D() with a 1x1 pixel works as expected and unbinds the buffer from the texture as documented here:
http://www.khronos.org/registry/gles/extensions/OES/OES_EGL_image.txt


BUG=264096

Review URL: https://chromiumcodereview.appspot.com/20245002