Move app shortcut manager into chrome/browser/apps
This code was previously in src/apps, which is meant to be independent
of chrome. The shortcut creation code is chrome specific so has been
moved to a new folder chrome/browser/apps, which is where chrome
specific apps non-ui code should now go.

BUG=159366

Review URL: https://chromiumcodereview.appspot.com/20286002