Allow supervised users to install themes.
Change the policy for supervised users to allow theme extensions.
Also allow resetting the theme to the default supervised user theme.
Hide the Use GTK+ button on Linux because the supervised user
theme replaces the default theme.
A screenshot of the settings page is attached to the bug description.

BUG=262937
TEST=Install a theme,and reset it on the settings page.

Review URL: https://chromiumcodereview.appspot.com/19774013