Fixes bug where child Widgets wouldn't get deactivated on aura.
This makes it so a native activation change is propagated to the
ActivationClient. This way child widgets see the activation change
(assuming they were active) and can update state. This means a child
bubble closes properly.

I'm holding off on the test for now as it requires an interactive ui
test.

BUG=262085
TEST=see bug
R=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/20150003