Get rid of chromium_default_blacklist().
Use a list of regexp instead, and create a function to add the --trace-blacklist
argument.

Add gen_blacklist() to precompile the regexes.

This is in preparation of reducing the number of references to chromium. It also
standardize the blacklists to be user configurable in all the scripts.

TBR=vadimsh@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/20746002