Prevent calling SetHideWebNotifications(false) when system notification is already destroyed.
This code arrives during constructor (ShelfWidget ctor -->
SystemTray::UpdateAfterLoginStateChanged --> UpdateNotifications)
and SetHideWebNotifications may invoke WebNotificationTray::ShowPopups
which registers an observer to ShelfLayoutManager, but it's still
in the construction of shelf, which is invalid.

This code is actually unnecessary when there are no system-tray
notifications, since web notification tray has ability to show its
popups when necessary.  The invocation of SetHideWebNotifications
will be only necessary when the destruction of system-tray notification
happened.

BUG=263664
R=oshima@chromium.org, stevenjb@chromium.org
TEST=new test case covers

Review URL: https://chromiumcodereview.appspot.com/20170002