Revert r212503, r214134, r214140.
All of the attempts to fix the link to the layout test results were broken.
Reverting them all until I can understand the code better and figure out
how to fix both the step-generating code and the log parsing.

The link will still be broken, but at least it'll be broken the way it
was before my changes.

R=iannucci@chromium.org
BUG=261732

Review URL: https://chromiumcodereview.appspot.com/20986002