Fix handling commands when executed with a shortcut.
Before, commands were not executed if they were formerly disabled, and later a shortcut was fired. This was happening because command state (enabled/disabled) is checked when focus changes.

This patch resolves this issue by rechecking if the command is enabled or disabled when firing a shortcut.

TEST=browser_test:*KeyboardOperations*
BUG=263703
R=nkostylev@chromium.org

Review URL: https://codereview.chromium.org/21039003