aura: Remove RefCountedIDMap.
The IDMap is owned by the GpuProcessTransportFactory,
and the destructor for the GpuProcTransportFactory asserts that
we have no per_compositor_data left.

ReflectorImpl references can be held by output surfaces, or by
per_compositor_data. The assert above clearly shows the latter
won't exist at shutdown.

For the former, per_compositor_data is destroyed in
RemoveCompositor() which happens in ~Compositor(). The
ui::Compositor class owns the LayerTreeHost, and the host_
is reset() before RemoveCompositor() happens. So this also
guarantees that there is no OutputSurface left when the
GpuProcessTransportFactory shuts down.

Therefore, there are no ReflectorImpls left, and they do not
need to hold a reference on the IDMap. Similarly, there are
also no output surfaces left, so they do not need to hold a
reference either. As such, we can remove the RefCountedIDMap
entirely, and just use an IDMap with raw pointers to it.

R=piman@chromium.org
BUG=258625

Review URL: https://codereview.chromium.org/20780002