Modern Download Manager browser action extension
Some of the features depend on https://codereview.chromium.org/16924017/
After that CL and this are submitted, I'll upload this to the webstore and start getting the By Google tag.

If you have a canary/dev/beta channel chrome installed or have a binary built from ToT and you're on the google network, go to
http://www.corp.google.com/~benjhayden/
right-click "download_manager.crx", Save Link As, open chrome://extensions and drag download_manager.crx from the shelf and drop it on chrome://extensions to install it.
If you install it on the stable channel, it will invite you to install the beta channel.

R=asanka@chromium.org, asargent@chromium.org, mkwst@chromium.org

Review URL: https://codereview.chromium.org/17286017