Add a main menu to the shim that matches the one in Chrome.
Currently the shim has no main menu. We construct one by adding a first
item (which gets replaced by OSX with the app's title) and hiding it.
Then a second item is added that will be used as the main menu.

BUG=168080
TEST=Start an app and close its window.
Cmd+Tab to the shim.
The main menu bar should have one unbolded item with the name of the app.

Review URL: https://chromiumcodereview.appspot.com/19490002