Implement ScrollEndEffectController
The class ScrollEndEffectController receives updates about the current
state of vertical overscroll and make adjustments to the layouts of the
elements of the browser to implement a scroll end effect. This CL adds
in a skelton of the class, hooks up the plumbing, etc. The effect itself
will be iterated on in follow on CLs.

BUG=25712
TEST=Built with logging in ScrollEndEffectController to confirm updates
     are being plumbed correctly.

Review URL: https://chromiumcodereview.appspot.com/19696011