Avoid losing focus on the search box in Settings.
Before, after clearing the search box, focus got moved to the first focusable element on the option page. Since the search box is outside of option pages, it caused releasing the focus, which is not the intended behaviour. This patch keeps the focus on the search box after clearing it.

To do so, focus is restored only, when previously another control was focused, but within the previous option page. Therefore, if there is another control focused, such as the search box or any other control outside of the option page, then it will not be released, which seem to be the intended behaviour.

TEST=Follow the bug's repro steps.
BUG=257326
R=dbeam@chromium.org, tkent@chromium.org

Review URL: https://codereview.chromium.org/19623004