Refactoring around entry changed event on FileCopyManager.
FileCopyManager had an event "copy-operation-complete". However, it is not
actually an event for operation complete, rather it is an event to notify
that some entries are changed. This CL renames the event, as well as callback
name in FileManager.
Also, entry change for 'delete' operation was notifyied in another way. The
changed entries' urls are attached to the 'deleted' event, and ButterBar
updated the metadata cache, thouth it is an UI component. By this CL,
"entry-changed" is also used for the entry change for 'delete' operation, too.

BUG=246976
TEST=Ran browser_tests --gtest_filter="*FileSystemExtensionApiTest*:*FileManagerBrowserTest*" and tested manually.
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/21092003