[password autofill] Don't autofill a form where any field can't be filled.
That is, if either the username or password field can't be filled (autocomplete="off", readonly, or disabled), don't fill the other field. Currently we will partially fill the form. This behavior is particularly surprising if the username can't be filled, as a password will be entered, but will disappear as soon as the user tries to enter a username.

BUG=264969
R=isherman@chromium.org

Review URL: https://codereview.chromium.org/20866002