Pepper Message Filters: Port to use explicit permission grants in ChildProcessSecurityPolicy.
This ports the Pepper message filters to use explicit permission grants instead of the base::PlatformFile bitmasks which are now deprecated. In the case that we have to deal with Pepper open flags, we now directly translate Pepper open flags to a series of explicit permissions checks instead of converting to base::PlatformFile bitmasks.

BUG=262142

Review URL: https://chromiumcodereview.appspot.com/19723010