Fix copy-paste in Files.app's share dialog.
Copying didn't work in the share dialog, since Ctrl+C was interpreted as a copying a selected file command. This patch resolves this issue by registering the copy and paste commands on the dialog container instead of the document. Note, that the share dialog is not a child of the dialog container, so when it is focused, then the commands will not be fired.

TEST=browser_tests, also fire the sharing dialog, try to copy the sharing url, and verify if it got copied to the clipboard.
BUG=265879
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/21325002