Revert WalletSigninHelperTest and WalletClientTest to TestingProfile instead of TestBrowserContext.
This is a components layer violation but the current code is exposing
us to a bad cast (see bug for details) and the autofill migration isn't
getting completed soon.  Reverting with a very narrow exception in the
DEPS file allowing these test to use TestingProfile.

BUG=266073

Review URL: https://chromiumcodereview.appspot.com/21113004