Pepper: Make DataFromWebURLResponse async.
Creating "new" FileRef resources in the renderer is an asynchronous operation.
Since DataFromWebURLResponse creates an external FileRef, that operation must
be asynchronous as well.

This requires that we add a small state machine to the PepperURLLoaderHost to
ensure that PpapiPluginMsg_URLLoader_ReceivedResponse is always sent to the
plugin before PpapiPluginMsg_URLLoader_FinishedLoading.

I've split this from the major FileRef refactor change in the interest of
making smaller patches (that are hopefully easier to review).

BUG=225441
R=bbudge@chromium.org, dmichael@chromium.org, yzshen@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=214384

Review URL: https://chromiumcodereview.appspot.com/20122004