Makes events outside display region of new style dialogs fall though.
https://codereview.chromium.org/20755006/ makes new style dialogs
display the right thing, but mouse events on parts of the dialog that
are not drawn still end up going to the dialog. This patch makes it so
that we end up setting a region on the HWND so that events can fall
through too.

BUG=241653
TEST=none
R=ben@chromium.org, msw@chromium.org

Review URL: https://chromiumcodereview.appspot.com/21145002