Do not load component extensions with background pages in perf tests.
Some Chrome component extensions have Javascript background pages or,
more often, event pages. On Chrome's first-run, or for a browser
upgrade, event pages are processed in a renderer process to find event
listeners. This allows events to be sent to that extension. The event
listeners are then stored in Preferences so that the event pages are not
processed on subsequent browser starts.

Component extensions are already skipped for most tests because they
interfere with regular testing, and the first-run event page
initialization does not reflect most common browser usage. Perf tests
already pass --no-first-run, but do not pass a --test-type argument, so
the existing check to skip component extensions was not hit.

This change adds --metrics-recording-only to the flags that indicate
component extensions should not be loaded. Previously, only --test-type
is checked.

Without this change, extra processes may start up in perf tests, which
interfere with the things the perf test is actually trying to measure.

BUG=251424
TEST=perf tests should more deterministic

Review URL: https://chromiumcodereview.appspot.com/21029002