Pepper: Hook up the in-process router for browser.
This is mostly stolen from raymes' change at:
  https://codereview.chromium.org/15774020/

but cleaned up to apply to master. It also adds in-process variants of the ResourceCall and ResourceReply messages, which are needed because the reply message from the browser to the renderer must be properly routed.

This change has been tested (along with a series of other changes) and was shown to work for in-process plugins in the "new" resource proxy.

TBR=jochen
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/19238007