Added CQ alerts for when logs are 120 seconds stale.
Heartbeat.py uploads status messages to chromelogs.appspot.com, which emails
alerts when a message with a status != 0 is sent.

This monitors /b/commit-queue/logs-chromium/commit_queue.log to see when its
last modified, and sets an erroneous status when its been modified more than 120
seconds ago.

Things that need to be done for this to work:
* A fresh version of pycrypto needs to be compiled and place next to heartbeat.py
* A cron job needs to call heartbeat.py -d every minute or so

BUG=255702

Review URL: https://chromiumcodereview.appspot.com/18493005