Pepper: Make PepperPluginInstance::Navigate async.
The primary motivation for this change is to enable the FileRef refactor; an
operation that Navigate depends on must be made asychronous. However, this
change cleans up some weirdness in PepperFlashRendererHost anyway, so I think
it's worth submitting as is so far. Plus, it makes sure this change is viable
before I go through modifying the whole call chain.

TBR=joi
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/21340006