Highlight items in the OSX App Launcher when install state changes.
When install begins, ensure the page is shown. When install completes,
ensure the page is shown, and select the item. This is to match CrOS/Win
behavior.

BUG=138633
TEST=With OSX App launcher enabled, install a large item from the
webstore. App launcher should show, with the item visible, with progress
bars. Move to a different page and dismiss the app launcher. When
install completes, app launcher should show again, and installed item
should be visible, and selected.

R=koz@chromium.org

Review URL: https://codereview.chromium.org/21458002