file_manager: Move ViewFilesFunction to private_api_tasks.h
Per discussion with yoshiki@, viewFiles() should belong to the tasks group, as
it does something similar to executeTask() We should eventually merge the two
functions, which would be a separate work item.

Along the way, rename ExecuteTasksFunction to ExecuteTaskFunction to be
consistent with the API function name.

BUG=257918
TEST=none
R=hashimoto@chromium.org

Review URL: https://codereview.chromium.org/21430004