Linux: use sandbox binary alongside chrome.
This removes linux_sandbox_path and instead expects sandbox binary
to be in the same directory as chrome (or CHROME_DEVEL_SANDBOX to be
set).

This will be helpful for SxS support on Linux, where the sandbox path
would be different depending on channel.

BUG=38598
R=avi@chromium.org, jln@chromium.org

Review URL: https://codereview.chromium.org/20197003