[rAC, OSX] Enable passing browser tests.
A few of the browser tests pass by now, as a side effect of work on
other bugs. Enabling them permanently, to increase test coverage.

This is a re-land of r214546 (https://codereview.chromium.org/19661005/),
which got reverted in r214587 (https://codereview.chromium.org/21088008/)

In addition to the enabled browser tests it also fixes a read-after-free.

The ASAN reported bug [1] is due to the fact that on Views/Aura,
closing a widget takes a round trip through the MessageLoop [2], while on OSX closing is immediate and OnConstrainedWindowClosed is called right after closing. Due to the callback structure of Dialog/Controller, that meant the dialog and controller would get deleted while still being in use further up the callstack.

[1]
http://build.chromium.org/p/chromium.memory/builders/Mac%20ASAN%20Tests%20%28...

[2] NativeWidgetAura::Close posts a CloseNow task, which does final cleanup.
https://code.google.com/p/chromium/codesearch#chromium/src/ui/views/widget/na...


BUG=256864

Review URL: https://chromiumcodereview.appspot.com/21006009