Authentication Manager for Google Now
There is currently no way to observe the sign-in state of the user in Chrome.
To work around this, the new authentication manager will poll for the
sign in state and notify any listeners about a change in the sign in state.

Calls for sign in state should correspondingly go through the authentication
manager since a check may already be in flight during a call to getAuthToken.

BUG=164227

Review URL: https://chromiumcodereview.appspot.com/21235008