file_manager: Simplify SetDriveMountPointPermissions() business
Previoulsy, SetDriveMountPointPermissions() was called in two places.
The call site of GetPreferencesFunction::RunImpl() looked very awkard.

Why did we need this? Looking at the revision history, this call site
was added in crrev.com/137102, so the function was called when the
Drive was enabled by the setting [1]. We can remove the call site
if we always call the function in RequestFileSystemFunction regardless
of whether or not the Drive is enabled in the settings, which is what
this patch does.

[1] This was tricky: when the Drive-enabled setting is changed,
onGDataPreferenceChanged event is sent to JavaScript, which calls
getGdataPreferences(), whicl calls GetGDataPreferencesFunction()
in C++, which calls AddGDataMountPoint(), which was later renamed
to SetDriveMountPointPermissions() in crrev.com/193705.

BUG=257918
TEST=Disable Drive from chrome:settings; Launch Files.app; Enable Drive from the settings page; Confirm that Drive in Files.app works just fine.
R=hashimoto@chromium.org, kinaba@chromium.org

Review URL: https://codereview.chromium.org/21476002