Cleanup: remove "tab" terminology from print preview code
With recent changes to web contents modal dialog management, the print
preview may be hosted on WebContents that are not browser tabs. Update
the names and comments to reflect this. This change was extracted from
r212329 which was reverted.

No functional difference is intended with this change.

Review URL: https://chromiumcodereview.appspot.com/21617002