Ensure --no-sandbox is specified if we're not using sandbox
This situation can occur if delegate->ShouldSandbox sets in_sandbox to false,
but the browser's command line does not have --no-sandbox.  In this situation,
the renderer thinks it should use a sandbox, but the sandbox has not been
initialized, causing the renderer to exit immediately.

BUG=

Review URL: https://chromiumcodereview.appspot.com/21030008