Provide minimal load timing info for GCF jobs.
This change introduces a GetLoadTimingInfo implementation for GCF's
URLRequestAutomationJob. The goal is not to provide meaningful load
timing, but rather to satisfy the URLRequestJob's contract. This
implemntation provides reasonable send_start and receive_headers_end
times from the standpoint of the job: these are the times that the
request is sent to IE's network stack and the time the headers are
returned. It also populates send_end with the same time as
receive_headers_end. This is the best we can do without changing the
interface for the automation channel.

BUG=265842
TEST=visit http://www.yahoo.com/ on XP/IE6 and hope for no Aw Snap!
R=robertshield@chromium.org, simonjam@chromium.org

Review URL: https://chromiumcodereview.appspot.com/21435003