Use the ProfileOAuth2TokenService at the UserCloudPolicyTokenForwarder.
This change replaces usage of the TokenService with the
ProfileOAuth2TokenService for the UserCloudPolicyTokenForwarder.

The UserCloudPolicyTokenForwarder now requests the access token from the
ProfileOAuth2TokenService and passes it to the UCPM, instead of passing the
refresh token. It has also been updated to wait for initialization of the
CloudPolicyService, to avoid minting access tokens when not necessary.

BUG=262130

Review URL: https://chromiumcodereview.appspot.com/20926003