Pepper: GetInfoForRenderer for N FileRefs.
Currently, operations to support AppendFileRefToBody in the renderer send a
sync message to the browser for each file ref.

This change alters GetInfoForRenderer (for the FileRef change) to take a vector
of FileRefs and return a vector of results back, reducing the number of IPC
round trips.

TBR=jschuh
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/21891002