Message center top-down UI fixes
Toasts in message center have been changed to move up when a notification is 
closed. Separator line in the button bar moved to the bottom in top down mode.
Few pixels of margin has been added to the message center work-area to prevent 
it from being flush with the edges of the screen.

BUG=267083

Review URL: https://chromiumcodereview.appspot.com/21626003