Don't let the user on a blank page when sign in is cancelled or fails.
Redirects to the NTP or apps page instead of leaving the user on a blank
page after sign in cancellation or failure. Before this CL, the user was
left on the blank page.

No redirection is necessary when:
- The sign in continue URL is not a blank page
  (i.e. sign in occured from an interstitial or a webstore installation).
- The blank page's tab is scheduled to be closed shortly after the failure
  (i.e. sign in occured from the Settings page).

TEST=
1. Make sure you are not signed in to Chromium.
2. Sign in with an enterprise account from:
- Startup screen, or
- NTP, or
- Hot-dog menu, or
- Extension installed bubble, or
- App launcher, or
- Apps page, or
- Bookmark bubble
** Check the "Choose what to sync checkbox". **
3. When you get the "You are signing in to a managed account" dialog, choose cancel.
Expected result: You are not left on a blank page.
BUG=238641

Review URL: https://chromiumcodereview.appspot.com/21109010