OfflineLoadPage: Get rid of code for displaying a link to mobile activation
This is legacy code that should have gone away with the offline page
update in crbug.com/170517. As it stands right now, the CSS is broken,
so we never show this option under any condition. Moreover, this code
is also wrong since we now have more than 1 mobile activation flow and 
this only works for Verizon 3G.

BUG=none
TEST=offline page should look the same as before

Review URL: https://chromiumcodereview.appspot.com/21806007