Deal with empty lists being omitted from the server's JSON response.
There are two places where we use lists in our REST API - the return of the List API (/hosts), and the tokenUrlPatterns list for each host.

Both places that handle those fields currently assume that an empty list will be returned explicitly as a 0-length array ([]) when in practice the fields are simply undefined. This CL fixes both places to deal with an undefined field correctly.

BUG=267293

Review URL: https://chromiumcodereview.appspot.com/21703002