Fixed loading resources from the host plugin's bundle on Mac.
The plugins shares the Cocoa class namespace with the Chrome's helper binary hosting the plugin. Because the client plugin is compiled into Chrome, Chrome's instance of NSBundleLocator is passed to NSBundle::bundleForClass, breaking resource loading as the result.

This CL switches to using dladdr() to get the bundle path instead of NSBUndle::bundleForClass.

BUG=267239

Review URL: https://chromiumcodereview.appspot.com/21976003