cc: Always have a LayerTreeHost in tests when we expect it in prod.
Many tests call functions on Layer without having the layer assigned
to a LayerTreeHost, but for functions that in production we require
a LayerTreeHost to be present. Add some DCHECKs in these methods that
a layer_tree_host_ is present, and make one present in the tests.

R=enne
BUG=267574
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/21856010