cc: Push valid property values when CalcDrawProps skips layer.
If CalcDropProps does not run on a layer for any reason, then we
do not SavePaintProperties() on the layer. When we push the layer's
properties, we can not push the paint properties since they are not
valid. We should instead push the raw value on the layer.

This is safe because if we didn't SavePaintProperties() we will also
not call Update() on the layer, so there is no chance for painting
to change any property values. This is enforced by a new DCHECK()
in Layer::Update() and we have all the subclasses of Layer call up
to the Update() method.

R=boliu, enne
BUG=267574

Review URL: https://chromiumcodereview.appspot.com/21839004