Recommit: Create a refcounted usb context wrapper
Create a refcounted usb context wrapper.
This will be also used when we introduce UsbDevice.

Fixed a thread check in UsbContext to make it work in unit tests.
Fixed the bug due to the fact that on Release build DCHECK is not evaluated.

Original CR https://codereview.chromium.org/20012002

TBR=asargent@chromium.org, pfeldman@chromium.org

BUG=223817

Review URL: https://chromiumcodereview.appspot.com/22104002