DevTools: Support workspace files indexing in browser process.
This patch introduces a trigram based index for file system folders added to DevTools.
Three methods are added to the interface between DevTools frontend and browser process:
 - IndexPath() performs indexing of the file system folder
 - SearchInPath() performs search in the index
 - StopIndexing() stops the indexing. This is needed because indexing is time consuming (seconds or even tens of seconds on huge folders).

The trigram based index is essentially a map from trigram to a list of files. The search method returns a list of files matching all trigrams that are present in the search query.

BUG=263313
R=pfeldman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/19851007