Fix extension IDL compiler to prevent incorrect code gen.
For properties of type "BINARY" (e.g. ArrayBuffer), compiler.py currently generates a call to "std::string(NULL)" in the data member list of type constructors. This call is invalid, as NULL is not a legal argument to std:string(). The fix is to not generate code at all in the case, so the default std::string constructor is called.

Also generate "kEventName" constant string for Events. This avoids having to manually define these constants in C++ API implementation code that needs to raise these events by name.

BUG=267656
BUG=268378

Review URL: https://chromiumcodereview.appspot.com/21944005