Browser Plugin: Plumb edit commands to the correct guest.
This CL addresses a bug where edit commands may go to the incorrect guest.
As a security measure, the browser process may kill the app that's trying
to forward an IPC to a guest it does not own. This may explain the crashes
seen in various webview apps.

BUG=251364
TBR=fsamuel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22258002