linux_aura: Implement the static part of the dbus menu for Unity.
In GTK, we get a menu in the Unity bar for free as long as we have a
GtkMenuBar object in our GtkWindow. linux_aura doesn't use gtk widgets
in the interface. So opportunistically dlopen libdbusmenu-glib and
manually construct a hidden DbusmenuMenuItem* object tree.

This only ports the static parts of the menus; it doesn't port the
History menu, which changes based on user actions.

BUG=265560

Review URL: https://chromiumcodereview.appspot.com/21187006