To dropp the decode done callback in RTCVideoDecoderBridgeTv.
WebRTC measures the decoding time and control the timing of the Decode()
calls. In the current implementation, we call decode done callback when
the browser process gets the frame. In order to get frames constantly
from WebRTC, this fix drops the decode done callback.
    
BUG=264585

Review URL: https://chromiumcodereview.appspot.com/20646003