Move the initialization code of volume item to VolumeItem constructor.
The initialization of volume item has been done in VolumeList.renderRoot_(), but as I wrote in the TODO comment in volume_list.js, it should be done in the constructor of VolumeItem. This patch moves it as the comment.

This patch shouldn't change any functionality.

BUG=none
TEST=Volume list works correctly on mounting/unmounting volume. Folder shortcut can be created and removed.
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/22021003