Add metrics for calculating precision/recall of link navigation pre-connect triggers.
This patch adds the UMA "Net.PreconnectTriggerUsed", and add the UMA "Net.PreconnectedLinkNavigation". 

PredictorTabHelper collects requests information for user link navigation. This information is sent to class Predictor::PreconnectUsage to see if the navigation has possibly used a preconnect session from {mouse,touch}-event triggers.

The UMA "Net.PreconnectTriggerUsed" records whether if a preconnect trigger is followed by a resource request (from link navigations) to the host or not. This is to measure precision of link-based preconnect triggers.

The UMA "Net.PreconnectedLinkNavigation" records whether if a resource request from link navigations is preceded by a preconnect trigger. This is to measure recall of link-based preconnect triggers.

BUG=235361

Review URL: https://chromiumcodereview.appspot.com/16514008