Make pepper revision choosable in nacl sdk update.
This will enable bots that require a nacl sdk to choose between stable, beta or canary. On the V8 waterfall, a builer is required that can choose something different then stable (will be an extra CL).

The naclsdk.zip is now also checked for changes. Before this CL, it always overwrote the up-to-date state of the installed bundles, which caused the bundles to get downloaded and installed in each build (even if they were up-to-date).

Review URL: https://chromiumcodereview.appspot.com/21623002