Settings: A text field in a list item should be focused when selection is moved by up/down cursor key.
Bug detail:
This is a regression caused by Blink r150870 [1].
When selection is moved by up/down cursor key, we set display:none to the
previously-focused INPUT element, then show another INPUT element and set focus
on it. The current code assumes elements with display:none keep focused, and
elements not in the list are never focused in this sequence. However it's a
wrong assumption, and Blink r150870 fixed the bug that unfocusable elements keep
focused.

Fix:
A list should not trigger focus cleanup code if any elemnts are focused
(activeElement is BODY.)

[1] http://src.chromium.org/viewvc/blink?revision=150870&view=revision

BUG=268291

Review URL: https://chromiumcodereview.appspot.com/22265004